refactor(*): move to ESM
